Add files via upload
One of the files is the pdf file which contains detailed analysis as well as our proposal to our instructor and project guide explaining what we aimed to achieve with this project and what business problems would we focus on. The other file is the tableau file which contains all our important and final dashboards along with the related sheets showing all our visualizations.